fix service tests
